Added gui
This commit contains the changes made in the cliet to respond to the servers messages.
A basic gui has been added. It does not yet respond to any input nor is it functional, however it can be rendered using the input that will be provided by the server at a later stage. Arbitrary games can be displayed as long as they fit within the restraints of 22x60 characters. Anything bigger will be cut off.
Added a display for the tictactoe game which can display arbitrary sizes, and render the insertion of new signs